Configurable method dispatch profiling
Method dispatch can now be enabled or disabled through `Opts`, and the
results of the profiling will either be saved to a file or directly dump
to the console depending on the value of `nativeProfileDispatch`.

If `nativeProfileDispatch` is set to `None`, then the results will be
displayed on the console. If a file is specified, then the results will
be written to the file.