Initial version of testing compiler
The testing compiler is made of two distinct components, as is the
scala-native compiler:
 - The testing NIRCompiler
 - The testing Optimizer

First, the testing NIRCompiler is used to compile scala code to NIR and
test the `nscplugin` project.

Second, the testing Optimizer is used to test the optimizer, and
ensuring that it performs the right transformations during its passes.