sbt test interface for Scala Native
The test interface works by stealing the tests from their original test
runner. This allows us to create a `main()` function that will run the
tests.

To add support for a new test framework, one needs to modify
`TestUtilities.scala` to add an entry explaining how to instantiate a
test for that particular framework.